added new cronjob that runs at 2AM every Sunday
